bg(db): Blocking unauthorized access (#52)
You can't review a session which doesn't exist
You can't review the same session twice
Admin only can change or delete
Should request session to only mentors and who exists
The same question shouldn't be send to the same Mentor.
you should be a mentor to accept/reject session and it should be addressed to you!

[##168426458]

+++ server/controllers/sessionController.js
@@ -5,21 +5,50 @@ import Db from '../db';
 
 const sessionController = {
   createSessions: async (req, res) => {
-    const data = await new Db().createSession(req.body, req.userData);
-    return Helper.handleSuccess(res, 201, 'Session request was sent, waiting mentor approval', data);
+    const {mentorId, questions} = req.body
+    const isMentor = await new Db().findByMultipleProp('users', 'userId', mentorId, 'role','mentor')
+    if (isMentor.length === 0) {
+      return Helper.handleError(res, 400, 'You are not allowed to request to this ID');
+    } else { 
+      const questionExist = await new Db().findByMultipleProp('sessions','mentorId', mentorId, 'questions',questions)
+      if(questionExist.length === 0 ) {
+        const data = await new Db().createSession(req.body, req.userData);
+        const {0: session} = data
+        return Helper.handleSuccess(res, 201, `Session request was sent, waiting mentor approval`, session);  
+      }
+
+        return Helper.handleError(res, 400, `You can't send the same question to the same mentor, please rephrase`);
+    }
   },
   getSessions: async (req, res) => {
     const { userId, role } = req.userData;
-    const data = await new Db().findByProp('sessions', 'menteeId', userId);
-    if (role === 'admin') {
-      const allSessions = await new Db().findAll('sessions');
-      return Helper.handleSuccess(res, 200, 'All Sessions', allSessions);
-    }
-    return Helper.handleSuccess(res, 200, 'Your sessions', data);
+      if (role === 'admin') {
+        const allSessions = await new Db().findAll('sessions');
+        if(allSessions.length === 0 ) {
+          return Helper.handleError(res, 404, 'There are no session yet');
+        }
+        return Helper.handleSuccess(res, 200, 'All Sessions', allSessions);
+      }
+      if(role === 'mentor'){
+        const data = await new Db().findByProp('sessions', 'mentorId', userId);
+        if(data.length === 0 ) {
+          return Helper.handleError(res, 404, 'You have no session yet');
+        }
+        return Helper.handleSuccess(res, 200, 'Your sessions', data);
+      }
+      const data = await new Db().findByProp('sessions', 'menteeId', userId);
+      return Helper.handleSuccess(res, 200, 'Your sessions', data);
+    
   },
 
   changeSessionStatus: async (req, res) => {
-    console.log(`${req.params.sessionId} I am here`)
+    const {userId:reqId, role} = req.userData
+    const sessionData = await new Db().findByProp('sessions', 'sessionId', req.params.sessionId)
+    const {0: sessionDetails} = sessionData
+    const {mentorid} = sessionDetails
+    if(role !== 'mentor' || reqId !== mentorid) {
+      return Helper.handleError(res, 400, 'You are not allowed to change the status of this session'); 
+    } else {
     const isPending = await new Db().findByMultipleProp('sessions', 'sessionid', parseInt(req.params.sessionId), 'status', 'pending');
     if (isPending.length === 0) {
       return Helper.handleError(res, 400, 'You are not allowed to change the status of this session');
@@ -33,21 +62,33 @@ const sessionController = {
     const data = {message: `Session is now Rejected`}
     return Helper.handleSuccess(res, 200, 'Success', data);
     }   
+    }
   },
   
   createReview: async (req, res) => {
+    const {userId, role} = req.userData
     const data = await new Db().createReview(req.body, req.params.sessionId, req.userData);
-    if(data === 'error') {
-      return Helper.handleError(res, 404, 'session not found');
-    } if(data === 'reviewed') {
-      return Helper.handleError(res, 400, 'You can\'t review again');
+    const sessionData = await new Db().findByProp('sessions', 'sessionId', req.params.sessionId)
+    const {0 : sessionDetails} = sessionData
+    const {menteeid} = sessionDetails
+    if (role !== 'user' || menteeid !== userId) {
+      return Helper.handleError(res, 400, `You are not allowed to review`);
+    } else {
+      if(data === 'error') {
+        return Helper.handleError(res, 404, 'session not found');
+      } if(data === 'reviewed') {
+        return Helper.handleError(res, 400, `You can't review again`);
+      }
+      return Helper.handleSuccess(res, 201, 'Thank you for reviewing. Kudos', data);
     }
-    return Helper.handleSuccess(res, 201, 'Thank you for reviewing. Kudos', data);
   },
 
   deleteReview: async (req, res) => {
-   const result = await new Db().deleteReview(req.params.sessionId);
-    return Helper.handleSuccess(res, 201, 'Review Succesfully Deleted');
+    if(req.userData.role === 'admin') {
+      const result = await new Db().deleteReview(req.params.sessionId);
+       return Helper.handleSuccess(res, 201, 'Review Succesfully Deleted');
+    }
+    return Helper.handleError(res, 401, 'Insufficient provilege. Please sign in');
   },
 
 };
